refactor oauth routes to use service
