+++ src/services/oauth/index.ts
@@ -1,0 +1,5 @@
+export { getApiOAuthService } from "./factory";
+export type {
+  OAuthService,
+  OAuthCallbackResult,
+} from "@/core/oauth/interfaces";
